pyproject.toml
[project]
name = "rag-form10q"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie>=1.1.2",
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "marker-pdf>=1.8.3",
    "openai>=1.99.3",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
